+++ client/src/pages/Calendar/Calendar.js
@@ -99,7 +99,7 @@ function Calendar() {
       {/* {!signedIn ? ( */}
       <div className="grid grid-cols-2 md:grid-cols-2 lg:grid-cols-2 2xl:grid-cols-2 justify-items-center gap-4">
         <div>
-          <h1 className="text-black sm:text-4xl text-3xl mb-1 font-medium title-font">
+          <h1 className="text-[#00df9a] sm:text-4xl text-3xl mb-1 font-medium title-font">
             JIM Calendar
           </h1>
           <GoogleLogin
@@ -119,18 +119,18 @@ function Calendar() {
           onSubmit={handleSubmit}
           className="lg:w-1/2 md:w-1/2 flex flex-col w-full "
         >
-          <h2 className="text-black sm:text-4xl text-3xl mb-1 font-medium title-font">
+          <h2 className="text-[#00df9a] sm:text-4xl text-3xl mb-1 font-medium title-font">
             Create an Event
           </h2>
           <div className="relative mb-4">
-            <label htmlFor="summary" className="leading-7 text-sm text-black">
+            <label htmlFor="summary" className="leading-7 text-sm text-[#00df9a]">
               Summary
             </label>
             <br />
             <input
               type="text"
               id="summary"
-              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-gray-100 py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
+              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-black py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
               value={summary}
               onChange={(e) => setSummary(e.target.value)}
             />
@@ -138,27 +138,27 @@ function Calendar() {
           <div className="relative mb-4">
             <label
               htmlFor="description"
-              className="leading-7 text-sm text-black"
+              className="leading-7 text-sm text-[#00df9a]"
             >
               Description
             </label>
             <br />
             <textarea
               id="description"
-              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-gray-100 py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
+              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-black py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
               value={description}
               onChange={(e) => setDescription(e.target.value)}
             />
           </div>
           <div className="relative mb-4">
-            <label htmlFor="location" className="leading-7 text-sm text-black">
+            <label htmlFor="location" className="leading-7 text-sm text-[#00df9a]">
               Location
             </label>
             <br />
             <input
               type="text"
               id="location"
-              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-gray-100 py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
+              className="w-full bg-gray-200 rounded border border-gray-700 focus:border-indigo-500 focus:ring-2 focus:ring-indigo-900 text-base outline-none text-black py-1 px-3 leading-8 transition-colors duration-200 ease-in-out"
               value={location}
               onChange={(e) => setLocation(e.target.value)}
             />
@@ -166,7 +166,7 @@ function Calendar() {
           <div className="relative mb-4">
             <label
               htmlFor="startDateTime"
-              className="leading-7 text-sm text-black"
+              className="leading-7 text-sm text-[#00df9a]"
             >
               Start Date
             </label>
@@ -182,7 +182,7 @@ function Calendar() {
           <div className="relative mb-4">
             <label
               htmlFor="endDateTime"
-              className="leading-7 text-sm text-black"
+              className="leading-7 text-sm text-[#00df9a]"
             >
               End Date
             </label>
@@ -197,7 +197,7 @@ function Calendar() {
           </div>
           <button
             type="submit"
-            className="text-white bg-green-800 border-0 py-2 px-6 focus:outline-none hover:bg-indigo-600 rounded text-lg"
+            className="text-white bg-[#00df9a] border-0 py-2 px-6 focus:outline-none hover:bg-indigo-600 rounded text-lg"
           >
             Submit Event
           </button>
