+++ client/src/pages/Contact/index.css
@@ -1,4 +1,77 @@
-@import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500&display=swap");
+.contact {
+  padding: 60px 0 200px 0;
+}
+.contact img {
+  width: 25%;
+}
+.contact h2 {
+	font-size: 20px;
+	font-weight: 700;
+  margin-bottom: 20px;
+  color: #00df9a;
+}
+.contact form input, .contact form textarea {
+  width: 50%;
+  background: rgba(172, 8, 8, 0.1);
+  border: 1px solid rgba(255, 255, 255, 0.5);
+  border-radius: 20px;
+  color: #fff;
+  margin: 0 0 8px 0;
+  padding: 18px 26px;
+  font-weight: 500;
+  font-size: 18px;
+  letter-spacing: 0.8px;
+  transition: 0.3s ease-in-out;
+}
+.contact form input:focus, .contact form textarea:focus {
+  background: rgba(255, 255, 255, 1);
+  color: #121212;
+}
+.contact form input::placeholder, .contact form textarea::placeholder {
+  font-size: 16px;
+  font-weight: 400;
+  color: #dcf634;
+}
+.contact form input:focus::placeholder, .contact form textarea:focus::placeholder {
+  color: #121212;
+  opacity: 0.8;
+}
+.contact form button {
+  font-weight: 700;
+  color: #000;
+  background-color: #fff;
+  padding: 14px 48px;
+  font-size: 18px;
+  margin-top: 25px;
+  border-radius: 0;
+  position: relative;
+  transition: 0.3s ease-in-out;
+}
+.contact form button span {
+  z-index: 1;
+  position: relative;
+}
+.contact form button:hover {
+  color: #fff;
+}
+.contact form button::before {
+  content: "";
+  background: #121212;
+  width: 0;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 0;
+  transition: 0.3s ease-in-out;
+}
+.contact form button:hover::before {
+  width: 50%;
+}
+
+
+
+/* @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@400;500&display=swap");
 
 .topcontainer {
   width: 100%;
@@ -80,4 +153,4 @@
   border-radius: 50%;
   font-size: 23px;
   margin: 0 20px;
-}
+} */
